requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
causal-conv1d==1.5.0.post8
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.1
coverage==7.8.0
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==2.19.2
decorator==5.2.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.1
evnn_pytorch==0.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fastrlock==0.8.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h5py==3.11.0
huggingface-hub==0.30.1
humanfriendly==10.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
hypothesis==6.97.6
identify==2.6.9
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
kenlm==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lightning-bolts==0.7.0
lightning-utilities==0.14.3
mamba-ssm==2.2.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mne==1.9.0
mne-bids==0.15.0
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
nltk==3.8.1
nodeenv==1.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.16.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
pre-commit==3.7.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==7.4.4
pytest-benchmark==4.0.0
pytest-cov==5.0.0
pytest-env==1.1.3
pytest-mock==3.14.0
pytest-rerunfailures==14.0
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytorch-lightning==1.8.6
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.10.1
sentry-sdk==2.25.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spikingjelly==0.0.0.0.14
submitit==1.5.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tokenizers==0.21.1
tomli==2.2.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==0.11.4
torchvision==0.18.0
tqdm==4.67.1
transformers==4.51.3
trio==0.23.2
triton==2.3.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.3.0
virtualenv==20.30.0
wandb==0.19.9
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
